+++ migrate/backup/main.go
@@ -2,10 +2,14 @@ package main
 
 import (
 	"encoding/json"
+	"fmt"
+	"io"
 	"os"
+	"strconv"
 	"strings"
 
 	did "github.com/ndidplatform/smart-contract/abci/did/v1"
+	"github.com/ndidplatform/smart-contract/migrate/utils"
 	"github.com/tendermint/iavl"
 	dbm "github.com/tendermint/tendermint/libs/db"
 )
@@ -15,19 +19,49 @@ var (
 )
 
 func main() {
+	// Variable
+	dbDir := getEnv("DB_NAME", "DB1")
+	dbName := "didDB"
+	backupDbDir := getEnv("BACKUP_DB_FILE", "Backup_DB")
+	backupDataFileName := getEnv("BACKUP_DATA_FILE", "data")
+	backupValidatorFileName := getEnv("BACKUP_VALIDATORS_FILE", "validators")
+	chainHistoryFileName := getEnv("CHAIN_HISTORY_FILE", "chain_history")
+	backupBlockNumberStr := getEnv("BLOCK_NUMBER", "")
+
 	// Delete backup file
-	fileName := "data"
-	deleteFile("migrate/data/" + fileName + ".txt")
+	deleteFile("migrate/data/" + backupDataFileName + ".txt")
+	deleteFile("migrate/data/" + backupValidatorFileName + ".txt")
+	os.Remove(backupDbDir)
+	deleteFile("migrate/data/" + chainHistoryFileName + ".txt")
 
-	validatorFileName := "validators"
-	deleteFile("migrate/data/" + validatorFileName + ".txt")
+	// Save previous chain info
+	resStatus := utils.GetTendermintStatus()
+	if backupBlockNumberStr == "" {
+		backupBlockNumberStr = resStatus.Result.SyncInfo.LatestBlockHeight
+	}
+	backupBlockNumber, err := strconv.ParseInt(backupBlockNumberStr, 10, 64)
+	if err != nil {
+		panic(err)
+	}
+	blockStatus := utils.GetBlockStatus(backupBlockNumber)
+	chainID := blockStatus.Result.Block.Header.ChainID
+	latestBlockHeight := blockStatus.Result.Block.Header.Height
+	latestBlockHash := blockStatus.Result.BlockMeta.BlockID.Hash
+	latestAppHash := blockStatus.Result.Block.Header.AppHash
+	fmt.Printf("--- Chain info at block: %s ---\n", backupBlockNumberStr)
+	fmt.Println("Chain ID: " + chainID)
+	fmt.Println("Latest Block Height: " + latestBlockHeight)
+	fmt.Println("Latest Block Hash: " + latestBlockHash)
+	fmt.Println("Latest App Hash: " + latestAppHash)
 
-	var dbDir = "DB1"
-	name := "didDB"
-	db := dbm.NewDB(name, "leveldb", dbDir)
+	// Copy stateDB dir
+	copyDir(dbDir, backupDbDir)
+
+	// Save kv from backup DB
+	db := dbm.NewDB(dbName, "leveldb", backupDbDir)
 	oldTree := iavl.NewMutableTree(db, 0)
 	oldTree.Load()
-	tree, _ := oldTree.GetImmutable(oldTree.Version())
+	tree, _ := oldTree.GetImmutable(backupBlockNumber)
 	_, ndidNodeID := tree.Get(prefixKey([]byte("MasterNDID")))
 	tree.Iterate(func(key []byte, value []byte) (stop bool) {
 		// Validator
@@ -39,10 +73,31 @@ func main() {
 			if err != nil {
 				panic(err)
 			}
-			fWriteLn(validatorFileName, jsonStr)
+			fWriteLn(backupValidatorFileName, jsonStr)
 			return false
 		}
-
+		// Chain history info
+		if strings.Contains(string(key), "ChainHistoryInfo") {
+			var chainHistory ChainHistory
+			if string(value) != "" {
+				err := json.Unmarshal([]byte(value), &chainHistory)
+				if err != nil {
+					panic(err)
+				}
+			}
+			var prevChain ChainHistoryDetail
+			prevChain.ChainID = chainID
+			prevChain.LatestBlockHeight = latestBlockHeight
+			prevChain.LatestBlockHash = latestBlockHash
+			prevChain.LatestAppHash = latestAppHash
+			chainHistory.Chains = append(chainHistory.Chains, prevChain)
+			chainHistoryStr, err := json.Marshal(chainHistory)
+			if err != nil {
+				panic(err)
+			}
+			fWriteLn(chainHistoryFileName, chainHistoryStr)
+			return false
+		}
 		if strings.Contains(string(key), string(ndidNodeID)) {
 			return false
 		}
@@ -59,9 +114,60 @@ func main() {
 		if err != nil {
 			panic(err)
 		}
-		fWriteLn(fileName, jsonStr)
+		fWriteLn(backupDataFileName, jsonStr)
 		return false
 	})
+}
+
+func copyDir(source string, dest string) (err error) {
+	sourceinfo, err := os.Stat(source)
+	if err != nil {
+		return err
+	}
+	err = os.MkdirAll(dest, sourceinfo.Mode())
+	if err != nil {
+		return err
+	}
+	directory, _ := os.Open(source)
+	objects, err := directory.Readdir(-1)
+	for _, obj := range objects {
+		sourcefilepointer := source + "/" + obj.Name()
+		destinationfilepointer := dest + "/" + obj.Name()
+		if obj.IsDir() {
+			err = copyDir(sourcefilepointer, destinationfilepointer)
+			if err != nil {
+				fmt.Println(err)
+			}
+		} else {
+			err = copyFile(sourcefilepointer, destinationfilepointer)
+			if err != nil {
+				fmt.Println(err)
+			}
+		}
+	}
+	return
+}
+
+func copyFile(source string, dest string) (err error) {
+	sourcefile, err := os.Open(source)
+	if err != nil {
+		return err
+	}
+	defer sourcefile.Close()
+	destfile, err := os.Create(dest)
+	if err != nil {
+		return err
+	}
+	defer destfile.Close()
+	_, err = io.Copy(destfile, sourcefile)
+	if err == nil {
+		sourceinfo, err := os.Stat(source)
+		if err != nil {
+			err = os.Chmod(dest, sourceinfo.Mode())
+		}
+
+	}
+	return
 }
 
 func prefixKey(key []byte) []byte {
@@ -103,4 +209,23 @@ func deleteFile(dir string) {
 	if err != nil {
 		panic(err)
 	}
+}
+
+type ChainHistoryDetail struct {
+	ChainID           string `json:"chain_id"`
+	LatestBlockHash   string `json:"latest_block_hash"`
+	LatestAppHash     string `json:"latest_app_hash"`
+	LatestBlockHeight string `json:"latest_block_height"`
+}
+
+type ChainHistory struct {
+	Chains []ChainHistoryDetail `json:"chains"`
+}
+
+func getEnv(key, defaultValue string) string {
+	value, exists := os.LookupEnv(key)
+	if !exists {
+		value = defaultValue
+	}
+	return value
 }