+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## TBD
+
+IMPROVEMENTS:
+
+- [DeliverTx] Add `chain_history_info` property to parameters of `InitNDID`.
+- [Query] Add new function (`GetChainHistory`).
+
 ## 0.11.2 (November 12, 2018)
 
 IMPROVEMENTS: